+++ tchannel/tchannel.py
@@ -79,6 +79,8 @@ class TChannel(object):
 
     """
 
+    FALLBACK = DeprecatedTChannel.FALLBACK
+
     def __init__(self, name, hostport=None, process_name=None,
                  known_peers=None, trace=False):
         """
@@ -235,6 +237,10 @@ class TChannel(object):
         return self._dep_tchannel.hostport
 
     def register(self, scheme, endpoint=None, handler=None, **kwargs):
+        if scheme is self.FALLBACK:
+            # scheme is not required for fallback endpoints
+            endpoint = scheme
+            scheme = None
 
         def decorator(fn):